refactor: extract escapeXml to shared module
Three byte-identical copies of `XML_ESCAPES` + `escapeXml` had accumulated
across `entity-disambiguator`, `reconcile-winner-pick`, and
`reconcile-pairwise-verify`. Promote the helper to a sibling module and
import it everywhere — the entity-disambiguator copy was the original, the
two reconcile-* copies arrived in this PR and should have re-used it from
day one.
