+++ src/lib/prompts/reconcile-pairwise-verify.ts
@@ -18,24 +18,13 @@
 import { z } from "zod";
 
 import type { CanonicalTopicKind } from "@/db/schema";
+import { escapeXml } from "@/lib/prompts/xml-escape";
 
 export interface ReconcileVerifySubject {
   id: number;
   label: string;
   kind: CanonicalTopicKind;
   summary: string;
-}
-
-const XML_ESCAPES: Record<string, string> = {
-  "&": "&amp;",
-  "<": "&lt;",
-  ">": "&gt;",
-  '"': "&quot;",
-  "'": "&apos;",
-};
-
-function escapeXml(s: string): string {
-  return s.replace(/[&<>"']/g, (ch) => XML_ESCAPES[ch] ?? ch);
 }
 
 function renderSubject(
